requirements.txt
discord.py
pycurl
parsedatetime
pytz
